Add Builder Log and other minor improvements

+++ OOOT GUI/Builder.cs
@@ -74,15 +74,18 @@ namespace OOOT_GUI
         /// Clone Repository
         public static bool Clone()
         {
+            Log.Message("Going to clone repository.");
+
             if (!Directory.Exists(InstallDir)) // create dir if it doesn't exist
             {
                 try
                 {
                     Directory.CreateDirectory(InstallDir);
+                    Log.Message("Created directory: " + InstallDir);
                 }
                 catch
                 {
-                    MessageBox.Show($"Error: Invalid installation directory. {InstallDir}", "Error!");
+                    ShowError($"Error: Invalid installation directory. {InstallDir}");
                     return false;
                 }
             }
@@ -90,7 +93,7 @@ namespace OOOT_GUI
             // directory exists, but is not empty, delete?
             if (Directory.Exists(GetOootPath()) && !DoesRepositoryExist())
             {
-                MessageBox.Show($"Can't clone because target directory is not empty! ({GetOootPath()})", "Error!");
+                ShowError($"Can't clone because target directory is not empty! ({GetOootPath()})");
 
                 // delete directory and continue?
                 if (!DeleteRepo())
@@ -101,12 +104,16 @@ namespace OOOT_GUI
             if (string.IsNullOrEmpty(CurrentBranch))
                 CurrentBranch = "master";
 
+            Log.Message("Cloning branch: " + CurrentBranch);
+
             // run command and get exit code
             int exitCode = CMD($"/C git clone --recursive -b {CurrentBranch} https://github.com/blawar/ooot.git", InstallDir);
 
             // process was aborted by user
             if (WasProcessAborted(exitCode))
             {
+                Log.Message("Cloning aborted by user.");
+
                 // delete downloaded files
                 if (!DoesRepositoryExist() && Directory.Exists(GetOootPath()))
                 {
@@ -120,29 +127,40 @@ namespace OOOT_GUI
                 return false;
             }
 
+            Log.Message("Cloned repository succesfully.");
+
             return true;
         }
 
         /// Git Pull Repository
         public static void Update()
         {
+            Log.Message("Going to update repository.");
+
             if (!DoesRepositoryExist())
             {
+                Log.Message("No repository found, aborting update.");
+
                 if (MessageBox.Show("No repository found. Do you want to clone it?", "Error!", MessageBoxButtons.YesNo) == DialogResult.Yes)
                     Clone();
                 return;
             }
 
             CMD("/C git pull", GetOootPath(), false);
+
+            Log.Message("Repository updated.");
         }
 
         /// Delete 'ooot' Folder
         public static bool DeleteRepo()
         {
             string oootPath = GetOootPath();
+
+            Log.Message("Going to delete repository from: " + oootPath);
+
             if (!Directory.Exists(oootPath))
             {
-                MessageBox.Show($"Can't delete repository, because directory not found! {oootPath})", "Error!");
+                ShowError($"Can't delete repository, because directory not found! {oootPath})");
                 return false;
             }
 
@@ -158,15 +176,18 @@ namespace OOOT_GUI
 
                 if (!Directory.Exists(oootPath)) // success
                 {
+                    Log.Message($"Repository deleted ({oootPath})");
                     MessageBox.Show("Repository deleted!");
                     return true;
                 }
                 else // failure
                 {
-                    MessageBox.Show("Something went wrong when deleting the repository!", "Error!");
+                    ShowError("Something went wrong when deleting the repository!");
                     return false;
                 }
             }
+
+            Log.Message("Cancel deleting repository.");
 
             return false;
         }
@@ -193,21 +214,43 @@ namespace OOOT_GUI
 
         public static void Build(bool isEurMqd)
         {
+            // check if VsBuildTools are still installing, before compiling
+            DialogResult result = IsVsBuildToolsStillInstalling();
+
+            if (result == DialogResult.Retry)
+            {
+                Log.Message("Vs Build Tools is still installing, can't compile yet.");
+                Build(isEurMqd);
+                return;
+            }
+            else if (result == DialogResult.Abort)
+            {
+                return;
+            }
+
+            Log.Message($"Going to build OOOT (Version: {GetRomVersion(isEurMqd)})");
+
             // check if repo exists
             if (!DoesRepositoryExist())
             {
+                Log.Message("Can't build, no repository found.");
+
                 // clone and continue building, or cancel and exit early
-                DialogResult result = MessageBox.Show("No repository found. Do you want to clone it?", "Error!", MessageBoxButtons.YesNo);
+                result = MessageBox.Show("No repository found. Do you want to clone it?", "Error!", MessageBoxButtons.YesNo);
                 if (result == DialogResult.Yes)
                     Clone();
                 else
                     return;
             }
 
+            Log.Message("Build: Checking for roms in ooot/roms folder...");
+
             // check if correct rom is present in 'ooot\roms' folder
             // if not, try to find and copy it from the Builder folder
             if (!IsRomInRomsFolder(isEurMqd))
             {
+                Log.Message("Build: No rom found from ooot/roms, trying to copy from builder folder.");
+
                 string romFileName = GetRomFilename(isEurMqd);
                 string romVersion = GetRomVersion(isEurMqd);
 
@@ -223,19 +266,29 @@ namespace OOOT_GUI
             if (WasProcessAborted(exitCode))
                 return; // user aborted building
 
-            if (System.IO.File.Exists(GetOootExePath())) // compiled OK, create shortcut?
+            if (System.IO.File.Exists(GetOootExePath())) // compiled OK, create shortcut and launch game?
             {
-                if (MessageBox.Show("OOOT compiled succesfully. Want to create shortcut?", "Finished", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                Log.Message("Built succesfully!");
+
+                result = MessageBox.Show("OOOT compiled succesfully. Want to create shortcut?", "Finished", MessageBoxButtons.YesNo);
+                if (result == DialogResult.Yes)
                     CreateShortcut();
+                else
+                {
+                    if (MessageBox.Show("Do you want to launch OpenOcarina?", "", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                        LaunchGame();
+                }
             }
             else // compile FAILED
             {
-                MessageBox.Show("Error: Something went wrong when building.", "Error!");
+                ShowError("Error: Something went wrong when building.");
             }
         }
 
         public static bool ExtractAssets(string romVersion)
         {
+            Log.Message($"Going to extract assets (Version: {romVersion})");
+
             if (!DoesRepositoryExist())
             {
                 // clone and continue extracting asset, or cancel and exit early
@@ -243,13 +296,18 @@ namespace OOOT_GUI
                 if (result == DialogResult.Yes)
                     Clone();
                 else
+                {
+                    Log.Message("Extracting Assets aborted because no repository found!");
                     return false;
+                }
             }
 
             // run command and get exit code
             int exitCode = CMD($"/C setup.py -c -b {romVersion}", GetOootPath(), true);
 
-            return WasProcessAborted(exitCode);
+            Log.Message("Extracting assets completed; setup.py returned exit code: " + exitCode);
+
+            return !WasProcessAborted(exitCode);
         }
 
         public static void LaunchGame()
@@ -268,14 +326,22 @@ namespace OOOT_GUI
                 MessageBox.Show($"No OOT.exe found! ({exePath})", "Error!");
         }
 
+        public static void ShowError(string message)
+        {
+            Log.Message("Error: " + message);
+            MessageBox.Show(message, "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+        }
+
         public static void CreateShortcut() // Create shortcut to OOOT
         {
+            Log.Message("Going to create shortcut to desktop.");
+
             string exePath = GetOootExePath();
 
             // show error and exit early, if no .exe found
             if (!System.IO.File.Exists(exePath))
             {
-                MessageBox.Show("Can't create shortcut, no OOT.exe found!", "Error!");
+                ShowError("Can't create shortcut, no OOT.exe found!");
                 return;
             }
 
@@ -293,17 +359,19 @@ namespace OOOT_GUI
 
                 if (System.IO.File.Exists(shortcutPath)) // succes, launch game?
                 {
+                    Log.Message("Desktop shortcut created.");
+
                     if (MessageBox.Show("Shortcut to OOOT created! Launch game?", "", MessageBoxButtons.YesNo) == DialogResult.Yes)
                         LaunchGame();
                 }
                 else // failure
                 {
-                    MessageBox.Show("Something went wrong, when trying to create shortcut!", "Error!");
+                    ShowError("Something went wrong, when trying to create shortcut!");
                 }
             }
             catch // even worse failure
             {
-                MessageBox.Show("ERROR: Something went wrong, when trying to create shortcut!", "Error!");
+                ShowError("ERROR: Something went wrong, when trying to create shortcut!");
             }
         }
 
@@ -312,21 +380,45 @@ namespace OOOT_GUI
         /// </summary>
         public static bool CopyRom(string filename, string romVersion, bool showErrorMessage = true)
         {
+            Log.Message($"Going to copy rom (Filename: {filename}, Version: {romVersion})");
+
+            // get path to directory containing roms
             string romDirPath = Path.Combine(GetOootPath(), @"roms\" + romVersion + @"\");
             if (!Directory.Exists(romDirPath))
             {
                 if (showErrorMessage)
-                    MessageBox.Show($"Can't copy ROM because OOOT roms directory is missing!", "Error!");
+                    ShowError($"Can't copy ROM because OOOT roms directory is missing!");
                 return false;
             }
 
             string source = Path.Combine(GetBuilderPath(), filename); // TODO: custom rom path
             string destination = Path.Combine(romDirPath + @"baserom_original.n64");
 
-            if (System.IO.File.Exists(source))
+            // rom is in builder folder, but not in ooot/roms; copy
+            if (System.IO.File.Exists(source) && !System.IO.File.Exists(destination))
+            {
+                Log.Message($"Copying ROM from {source} to {destination}");
                 System.IO.File.Copy(source, destination, true);
+            }
+            // rom is already in ooot/roms; no need to copy
+            else if (System.IO.File.Exists(destination))
+            {
+                Log.Message($"Rom already exist in destination: {destination}");
+                return true;
+            }
+            else
+            {
+                Log.Message($"No ROM found from the Builder folder ({source})");
+            }
 
-            return System.IO.File.Exists(source);
+            bool result = System.IO.File.Exists(destination);
+
+            if (result)
+                Log.Message($"Rom found from: {destination}");
+            else
+                Log.Message($"No ROM found from: {destination}");
+
+            return result;
         }
 
         /// <summary>
@@ -348,6 +440,8 @@ namespace OOOT_GUI
         /// </summary>
         public static bool DownloadTools(bool installTools, bool showMessages, bool ignoreInstalledCheck = false)
         {
+            Log.Message("Going to download tools.");
+
             bool git = false;
             bool python = false;
             bool vsBuild = false;
@@ -362,7 +456,10 @@ namespace OOOT_GUI
                 if (git && python && vsBuild)
                 {
                     if (showMessages)
+                    {
                         MessageBox.Show("All tools are already installed.");
+                        Log.Message("Tools are already installed.");
+                    }
                     return true;
                 }
             }
@@ -370,7 +467,7 @@ namespace OOOT_GUI
             string tmpDir = GetTempDownloadPath(true);
             if (!Directory.Exists(tmpDir))
             {
-                MessageBox.Show($"Can't create temporary download directory for tools! ({tmpDir}", "Error!");
+                ShowError($"Can't create temporary download directory for tools! ({tmpDir}");
                 return false;
             }
 
@@ -407,7 +504,7 @@ namespace OOOT_GUI
             // execute command(s)
             int exitCode = CMD(command, tmpDir, true);
             if (WasProcessAborted(exitCode))
-                return false; // use canceled download, exit.
+                return false; // user canceled download, exit.
 
             // install tool after download, if set so
             if (installTools)
@@ -421,6 +518,8 @@ namespace OOOT_GUI
         /// </summary>
         public static bool InstallTools(bool showMessages)
         {
+            Log.Message("Going to install tools.");
+
             bool git = IsGitInstalled();
             bool python = IsPythonInstalled();
             bool vsBuild = IsVsBuildToolsInstalled();
@@ -428,6 +527,7 @@ namespace OOOT_GUI
             // tools are already installed, show optional message and early exit
             if (git && python && vsBuild)
             {
+                Log.Message("Tools already installed.");
                 if (showMessages)
                     MessageBox.Show("All tools are already installed.");
                 return true;
@@ -437,7 +537,7 @@ namespace OOOT_GUI
             string tmpDir = GetTempDownloadPath(false);
             if (!Directory.Exists(tmpDir))
             {
-                MessageBox.Show($"Can't find temporary download directory for tools! ({tmpDir}", "Error!");
+                ShowError($"Can't find temporary download directory for tools! ({tmpDir}");
                 return false;
             }
 
@@ -480,23 +580,26 @@ namespace OOOT_GUI
             string htsFilename = "THE LEGEND OF ZELDA_HIRESTEXTURES.hts";
             string htsPath = Path.Combine(oootReleasePath, htsFilename);
 
+            Log.Message("Going to download texture pack from: " + texturePackUrl);
+
             // no release folder, exit
-            if (!System.IO.Directory.Exists(oootReleasePath))
+            if (!Directory.Exists(oootReleasePath))
             {
-                MessageBox.Show($"No OOOT 'Release' folder found! ({oootReleasePath})", "Error!");
+                ShowError($"No OOOT 'Release' folder found! ({oootReleasePath})");
                 return;
             }
 
             // no 7zr.exe, exit
             if (!System.IO.File.Exists(toolPath))
             {
-                MessageBox.Show($"No '7zr.exe' found from 'Builder/Tools' folder! ({toolPath})", "Error!");
+                ShowError($"No '7zr.exe' found from 'Builder/Tools' folder! ({toolPath})");
                 return;
             }
 
             // texture pack is already installed, exit
             if (System.IO.File.Exists(htsPath))
             {
+                Log.Message("Texture pack is already installed.");
                 MessageBox.Show("Texture pack is already installed!");
                 return;
             }
@@ -509,10 +612,11 @@ namespace OOOT_GUI
                     return; // user aborted downloading
             }
 
-
             // extract texture pack
             if (System.IO.File.Exists(texturePackPath))
             {
+                Log.Message("Texture pack download. Going to extract.");
+
                 Process p = new Process();
                 p.StartInfo.FileName = toolPath;
                 p.StartInfo.Arguments = $"e {texturePackPath}";
@@ -525,6 +629,8 @@ namespace OOOT_GUI
                 int exitCode = p.ExitCode;
                 if (WasProcessAborted(exitCode))
                 {
+                    Log.Message("Texture pack extracting aborted.");
+
                     // delete .7zip file
                     if (System.IO.File.Exists(texturePackPath))
                         System.IO.File.Delete(texturePackPath);
@@ -534,6 +640,8 @@ namespace OOOT_GUI
 
                 if (System.IO.File.Exists(htsPath)) // installed texture pack succesfully
                 {
+                    Log.Message("Texture pack installed.");
+
                     // delete .7zip file?
                     if (MessageBox.Show($"Texture pack installed! Do you want to delete '{texturePackFilename}'?", "Completed", MessageBoxButtons.YesNo) == DialogResult.Yes)
                     {
@@ -544,19 +652,20 @@ namespace OOOT_GUI
                         }
                         catch
                         {
-                            MessageBox.Show($"Error when deleting {texturePackPath}!", "Error!");
+                            ShowError($"Error when deleting {texturePackPath}!");
                         }
                     }
                 }
                 else // failed texture pack install
                 {
-                    MessageBox.Show($"No file found! ({texturePackPath})", "Error!");
+                    ShowError($"No file found! ({texturePackPath})");
                 }
             }
         }
 
         public static int CMD(string command, string workingDir = "", bool showWindow = true)
         {
+            Log.Message($"CMD: {command} (WorkingDir: {workingDir})");
             Process p = new Process();
             p.StartInfo.FileName = "CMD.exe";
             p.StartInfo.Arguments = command;
@@ -617,6 +726,30 @@ namespace OOOT_GUI
         public static bool IsAllToolsInstalled()
         {
             return IsGitInstalled() && IsPythonInstalled() && IsVsBuildToolsInstalled();
+        }
+
+        public static DialogResult IsVsBuildToolsStillInstalling()
+        {
+            bool block = false;
+            Process[] vsProcesses = Process.GetProcesses();
+            foreach (Process p in vsProcesses)
+            {
+                if (p.MainWindowTitle.Contains("Visual Studio Installer"))
+                    block = true;
+                else if (p.ProcessName == "vs_BuildTools")
+                    block = true;
+                else if (p.ProcessName == "vs_setup_bootstrapper")
+                    block = true;
+            }
+
+            // vs build tools processes detected, can't compile yet
+            if (block)
+            {
+                DialogResult result = MessageBox.Show("Wait for VS Build Tools to install before compiling.", "Error!", MessageBoxButtons.AbortRetryIgnore);
+                return result;
+            }
+
+            return DialogResult.Ignore;
         }
 
         private static string GetCmdOutput(string command, string workingDir = "")
@@ -728,7 +861,10 @@ namespace OOOT_GUI
         {
             bool result = (exitCode == -1073741510);
             if (result)
+            {
+                Log.Message("Process aborted by user.");
                 MessageBox.Show("Process aborted!");
+            }
             return result;
         }
